调试ThinkOut流程;

+++ SMG_NothingIsAll/AIFoundation/AIClass/Utils/SMGUtils.m
@@ -62,7 +62,7 @@ static int readDiskCount;
 //Direction的mv分区pointer;(存引用序列)
 +(AIKVPointer*) createPointerForDirection:(NSString*)mvAlgsType direction:(MVDirection)direction{
     NSInteger pointerId = 0;
-    AIKVPointer *kvPointer = [AIKVPointer newWithPointerId:pointerId folderName:kPN_DIRECTION(direction) algsType:mvAlgsType dataSource:DefaultDataSource isOut:false isMem:false];
+    AIKVPointer *kvPointer = [AIKVPointer newWithPointerId:pointerId folderName:kPN_DIRECTION((long)direction) algsType:mvAlgsType dataSource:DefaultDataSource isOut:false isMem:false];
     return kvPointer;
 }
